Remove dependency on Apache Commons IO.
Split out result wrapping/unwrapping code into separate class
RedisResult.
